requirements.txt
requests
Beautifulsoup4
django-allauth